cmd: fastboot: allow fastboot from arbitrary USB port
The CONFIG_FASTBOOT_USB_DEV is unconditionally defined whenever
the fastboot feature is enabled.

When defined, the configuration parameter discards the interface
index specified in the command line, unless 'auto' option is used.

This patch treats CONFIG_FASTBOOT_USB_DEV=-1 as if the default
fast boot device is not specified thus allowing to use fastboot
command for arbitrary USB interface.

Signed-off-by: Dmitry Bazhenov <dimab@voxelbotics.com>